requirements.txt
av
einops
h5py
matplotlib
numpy
Pillow
pytorch-ignite
pytorch-lightning
lightning-flash
wandb
torch
six
torchvision
decord

